BIG refactor of algorithms and tests
