+++ athletic_pandas/algorithms/critical_power.py
@@ -3,6 +3,10 @@ from lmfit import Parameters, minimize
 
 
 def _ecp_predict_5_3(model_params, x):
+    """
+    Credits to Damien Grauser. Source:
+    https://github.com/GoldenCheetah/GoldenCheetah/blob/master/src/Metrics/ExtendedCriticalPower.cpp
+    """
     model = (
         (
             (
@@ -35,6 +39,10 @@ def _ecp_predict_5_3(model_params, x):
 
 
 def _ecp_predict_7_3(model_params, x):
+    """
+    Credits to Damien Grauser. Source:
+    https://github.com/GoldenCheetah/GoldenCheetah/blob/master/src/Metrics/ExtendedCriticalPower.cpp
+    """
     model = (
         (
             (
@@ -74,11 +82,7 @@ def _ecp_residual(model_params, x, data, version):
     return data - model
 
 
-def extended_critical_power(mean_max_power, **kwargs):
-    """
-    Credits to Damien Grauser. Source:
-    https://github.com/GoldenCheetah/GoldenCheetah/blob/master/src/Metrics/ExtendedCriticalPower.cpp
-    """
+def critical_power_model(mean_max_power, version='5_3', initial_model_params=None):
     mean_max_power = mean_max_power[:5400]
 
     time_axis = np.asarray(range(1, len(mean_max_power) + 1))
@@ -100,6 +104,4 @@ def extended_critical_power(mean_max_power, **kwargs):
 
     model = minimize(_ecp_residual, model_params, args=(time_axis, mean_max_power, version))
 
-    predictions = _ecp_predict_5_3(model.params, time_axis)
-
-    return predictions
+    return model
