+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import pandas as pd
+import pytest
+
+
+@pytest.fixture
+def power():
+    return pd.Series(range(100))
